feat: add database_connect

+++ webservice/src/models.rs
@@ -4,8 +4,8 @@ use serde::{Deserialize, Serialize};
 
 #[derive(Debug, Clone, Deserialize, Serialize)]
 pub struct Course {
-    pub teacher_id: usize,
-    pub id: Option<usize>,
+    pub teacher_id: i32,
+    pub id: Option<i32>,
     pub name: String,
     pub time: Option<NaiveDateTime>,
 }
